Common clone_deserialize_error
